# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PetMon API
  description: Pet monitoring and nutrition tracking API
  version: 1.0.0
  contact:
    name: PetMon Team
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Main API server
security:
  - bearerAuth: []
paths:
...
  /pets:
    get:
      tags:
        - Pets
      summary: List all pets
      x-required-scope: api_read
      description: Retrieve a list of all pets in the system
      responses:
        '200':
          description: List of pets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pet'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Pets
      summary: Create a new pet
      x-required-scope: api_write
      description: Create a new pet record with the provided information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePet'
      responses:
        '201':
          description: Pet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /pets/{id}:
    get:
      tags:
        - Pets
      summary: Get a pet by ID
      x-required-scope: api_read
      description: Retrieve a specific pet by its UUID
      parameters:
        - name: id
          in: path
          required: true
          description: Pet UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pet information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Pets
      summary: Update a pet
      x-required-scope: api_write
      description: Update one or more fields of an existing pet
      parameters:
        - name: id
          in: path
          required: true
          description: Pet UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePet'
      responses:
        '200':
          description: Pet updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Pets
      summary: Delete a pet
      x-required-scope: api_write
      description: Delete a pet by its UUID
      parameters:
        - name: id
          in: path
          required: true
          description: Pet UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Pet deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /nutrition/schedules:
    get:
      tags:
        - Nutrition Schedules
      summary: List nutrition schedules
      x-required-scope: api_read
      description: Retrieve nutrition schedules for all pets or a specific pet
      parameters:
        - name: pet_id
          in: query
          required: false
          description: Filter by pet UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of nutrition schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NutritionSchedule'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Nutrition Schedules
      summary: Create nutrition schedule
      x-required-scope: api_write
      description: Create a new nutrition schedule for a pet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNutritionSchedule'
      responses:
        '201':
          description: Nutrition schedule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionSchedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /nutrition/schedules/{id}:
    get:
      tags:
        - Nutrition Schedules
      summary: Get nutrition schedule by ID
      x-required-scope: api_read
      description: Retrieve a specific nutrition schedule
      parameters:
        - name: id
          in: path
          required: true
          description: Schedule ID
          schema:
            type: string
      responses:
        '200':
          description: Nutrition schedule
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionSchedule'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags:
        - Nutrition Schedules
      summary: Update nutrition schedule
      x-required-scope: api_write
      description: Update one or more fields of a nutrition schedule
      parameters:
        - name: id
          in: path
          required: true
          description: Schedule ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNutritionSchedule'
      responses:
        '200':
          description: Nutrition schedule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NutritionSchedule'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Nutrition Schedules
      summary: Delete nutrition schedule
      x-required-scope: api_write
      description: Delete a nutrition schedule
      parameters:
        - name: id
          in: path
          required: true
          description: Schedule ID
          schema:
            type: string
      responses:
        '204':
          description: Nutrition schedule deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer authentication with either:
        - OIDC JWT token (format: standard JWT) — full access, no scope restrictions
        - API token with `pm_api_` prefix — access controlled by the token's scope set

        **API token scopes:**
        | Scope | Access granted |
        |---|---|
        | `all` | Full access — equivalent to OIDC |
        | `api_read` | All GET endpoints |
        | `api_write` | All POST / PATCH / DELETE endpoints |
        | `mcp` | MCP tool endpoint (`/mcp`) |

        A token may hold multiple scopes (e.g. `["api_read", "mcp"]`).
        The scope required by each endpoint is shown in the `x-required-scope` field.
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Token does not have the required scope for this endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    PetSpecies:
      type: string
      enum:
        - cat
        - dog
        - bunny
        - parrot
        - other
      description: Pet species
    PetStatus:
      type: string
      enum:
        - active
        - archived
      description: Pet status
    Pet:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Pet UUID
        name:
          type: string
          description: Pet name
        species:
          $ref: '#/components/schemas/PetSpecies'
        status:
          $ref: '#/components/schemas/PetStatus'
        breed:
          type: string
          nullable: true
          description: Pet breed
        birth_date:
          type: string
          nullable: true
          description: Pet birth date (YYYY-MM-DD)
        blood_type:
          type: string
          nullable: true
          description: Pet blood type
        color:
          type: string
          nullable: true
          description: Pet color/markings
        feeding_notes:
          type: string
          nullable: true
          description: Feeding instructions and notes
        telegram_chat_id:
          type: string
          nullable: true
          description: Telegram chat ID for notifications
        telegram_thread_id:
          type: string
          nullable: true
          description: Telegram thread ID for topic replies
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp (RFC3339)
        updated_at:
          type: string
          format: date-time
          description: Record last update timestamp (RFC3339)
      required:
        - id
        - name
        - species
        - status
        - created_at
        - updated_at
    CreatePet:
      type: object
      properties:
        name:
          type: string
          description: Pet name
        species:
          $ref: '#/components/schemas/PetSpecies'
        status:
          $ref: '#/components/schemas/PetStatus'
        breed:
          type: string
          nullable: true
        birth_date:
          type: string
          nullable: true
          format: date
        blood_type:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        feeding_notes:
          type: string
          nullable: true
        telegram_chat_id:
          type: string
          nullable: true
        telegram_thread_id:
          type: string
          nullable: true
      required:
        - name
    UpdatePet:
      type: object
      properties:
        name:
          type: string
          nullable: true
        species:
          $ref: '#/components/schemas/PetSpecies'
        status:
          $ref: '#/components/schemas/PetStatus'
        breed:
          type: string
          nullable: true
        birth_date:
          type: string
          nullable: true
          format: date
        blood_type:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        feeding_notes:
          type: string
          nullable: true
        telegram_chat_id:
          type: string
          nullable: true
        telegram_thread_id:
          type: string
          nullable: true
...
    NutritionScheduleRule:
      type: object
      properties:
        category:
          type: string
          description: Nutrition category
        target_amount:
          type: number
          format: double
          description: Target amount
        unit:
          type: string
          nullable: true
          description: Unit of measurement
        time_of_day:
          type: string
          nullable: true
          description: Recommended time (e.g., HH:MM)
        notes:
          type: string
          nullable: true
          description: Additional notes
    NutritionSchedule:
      type: object
      properties:
        id:
          type: string
          description: Schedule ID
        pet_id:
          type: string
          format: uuid
          description: Pet UUID
        name:
          type: string
          description: Schedule name
        active:
          type: boolean
          description: Whether the schedule is active
        rules_json:
          type: string
          description: JSON-serialized array of NutritionScheduleRule objects
        created_at:
          type: string
          format: date-time
          description: Record creation timestamp (RFC3339)
        updated_at:
          type: string
          format: date-time
          description: Record last update timestamp (RFC3339)
      required:
        - id
        - pet_id
        - name
        - active
        - rules_json
        - created_at
        - updated_at
    CreateNutritionSchedule:
      type: object
      properties:
        pet_id:
          type: string
          format: uuid
          description: Pet UUID
        name:
          type: string
          description: Schedule name
        active:
          type: boolean
          nullable: true
          description: Active status (defaults to true)
        rules:
          type: array
          nullable: true
          description: Array of schedule rules
          items:
            $ref: '#/components/schemas/NutritionScheduleRule'
      required:
        - pet_id
        - name
    UpdateNutritionSchedule:
      type: object
      properties:
        name:
          type: string
          nullable: true
        active:
          type: boolean
          nullable: true
        rules:
          type: array
          nullable: true
          description: Array of schedule rules
          items:
            $ref: '#/components/schemas/NutritionScheduleRule'
...